Merged revisions 114597 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r114597 | russell | 2008-04-23 15:49:18 -0500 (Wed, 23 Apr 2008) | 10 lines

Fix an issue that caused getting the correct next channel to not always work.
Also, remove setting the amount of time to wait for a digit from 5 seconds back
down to 1/10 of a second.  I believe this was so the beep didn't get played over
and over really fast, but a while back I put in another fix for that issue.

(closes issue #12498)
Reported by: jsmith
Patches:
      app_chanspy_channel_walk.trunk.patch uploaded by jsmith (license 15)

........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@114598 f38db490-d61c-443f-a65b-d21fe96a405b