Merge code from team/russell/parking_updates
Add some additional features to the core park_call_full() function, and expose
them as options to the Park() application.  The functionality being added is the
ability to specify a custom return extension/context/priority, a custom timeout,
and a couple of options.  The options are to play ringing instead of MOH to the
parked caller, and to randomize parking spot selection.

(code inspired by the patch in AST-17, code from switchvox)


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@114655 f38db490-d61c-443f-a65b-d21fe96a405b