(closes issue #13041) Reported by: eliel
OK, now the context registrar slot is strdup'd. It is freed
on destruction. I don't see the need to do this with all
the structs' registrar fields, but if some wild case proves
they should also be handled this way, then we can 
put in the extra work at that time.




git-svn-id: http://svn.digium.com/svn/asterisk/trunk@130297 f38db490-d61c-443f-a65b-d21fe96a405b