Merged revisions 132826 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r132826 | crichter | 2008-07-23 13:37:50 +0200 (Mi, 23 Jul 2008) | 1 line

another Fix because of r119585, this commit has broken high frequented BRI Ports, there was a possibility that a channel, that was marked as in_use would be reused later, the corresponding port could got stuck then. So it is recommended to upgrade for chan_misdn users.
........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@132883 f38db490-d61c-443f-a65b-d21fe96a405b