When determining if codecs used by SIP peers allow the media to be natively bridged, use the jointcapability instead of the peercapability.
It seems that the intent of using the peercapability was to
expand the choice of codecs for the call to increase the
chances of being able to native bridge the channels. The 
problem is that if a codec were settled on for the native
bridge and that wasn't a codec that was configured to be used
by Asterisk for that peer, then Asterisk would send a 
REINVITE with no codecs in the SDP which is a bug no matter
how you slice it.


(closes issue #13076)
Reported by: ramonpeek
Patches:
      13076.patch uploaded by putnopvut (license 60)
Tested by: tbelder



git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@142079 f38db490-d61c-443f-a65b-d21fe96a405b