Blocked revisions 162670 via svnmerge
........
r162670 | mmichelson | 2008-12-10 10:44:37 -0600 (Wed, 10 Dec 2008) | 14 lines

Update to stringfield handling so that side-effects on
parameters are not evaluated multiple times.

An example where this caused a problem was in chan_sip.c, with
the line

  ast_string_field_set(p, fromdomain, ++fromdomain);

This patch was originally uploaded to issue #13783 by
jamessan. While the issue was closed for other reasons, this
patch is valid and fixes a separate problem, and is thus
being committed.


........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@162672 f38db490-d61c-443f-a65b-d21fe96a405b