Update extensions.conf.sample to be correct.
In trunk, the only necessary change pointed out was that the call
to ChanIsAvail uses an option that has been removed.

For the 1.6.1 branch, however, it appears that the sample file is
badly in need of updating since there are |'s used all over the place
there. My tentative plan is just to copy trunk's sample config file
to those branches since the info there is most up-to-date and should
be correct for use in 1.6.1

Thanks to macli in #asterisk-dev for bringing this up



git-svn-id: http://svn.digium.com/svn/asterisk/trunk@173776 f38db490-d61c-443f-a65b-d21fe96a405b