Allow for "magic" pickups to work when we wish to ignore the context
When the subscription context for a call pickup subscription differs
from the context of the call pickup target, there's not an easy way
to divine what context should be used for the pickup. The way to work
around this is to use PICKUPMARK as the context for the pickup.

This has been documented in the sip.conf.sample file

(ABE-1708)



git-svn-id: http://svn.digium.com/svn/asterisk/trunk@180155 f38db490-d61c-443f-a65b-d21fe96a405b