Fix errors in cdr_custom that cause reference errors when non-CDR variable substitution is done.
cdr_custom was creating a ast_channel struct directly and passing it into the
core for variable substition.  This was fine as long as the format string
contained only calls to the CDR() function.  Doing something like ${EPOCH} on
the other hand tried to lock the channel, which would fail and throw an error
because the passed channel hadn't been allocated as an ao2 object.  So now we
create the dummy channel with ast_channel_alloc, and everything works as
expected.


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@196520 f38db490-d61c-443f-a65b-d21fe96a405b