Add ActionID to CoreShowChannel event.
There is inconsistency in how we handle manager responses that are lists of
items and, unfortunately, third parties have come to rely on ActionID being on
every event within those lists instead of just keeping track of the ActionID for
the current response.  This change makes CoreShowChannels include the ActionID
with each CoreShowChannel event generated as a result of it being called.

(closes issue #15001)
Reported by: sum
Patches:
      patchactionid2.patch uploaded by sum (license 766)


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@196945 f38db490-d61c-443f-a65b-d21fe96a405b