Ensure that realtime mailboxes properly report status on subscription.
This patch modifies app_voicemail's response to mailbox status subscriptions
(via the internal event system) to ensure that a subscription triggers an
explicit poll of the mailbox, so the subscriber can get an immediate cached
event with that status. Previously, the cache was only populated with the
status of non-realtime mailboxes.

(closes issue #15717)
Reported by: natmlt



git-svn-id: http://svn.digium.com/svn/asterisk/trunk@213697 f38db490-d61c-443f-a65b-d21fe96a405b