port string to int conversion using sscanf
There are several instances where a port is parsed
from a uri or some other source and converted to
an int value using atoi(), if for some reason the
port string is empty, then a standard port is used.
This logic is used over and over, so I created a function
to handle it in a safer way using sscanf().



git-svn-id: http://svn.digium.com/svn/asterisk/trunk@215681 f38db490-d61c-443f-a65b-d21fe96a405b