Blocked revisions 217668 via svnmerge
........
r217668 | oej | 2009-09-10 21:07:24 +0200 (Tor, 10 Sep 2009) | 9 lines

Remove harmful code that causes endless loops. 
Remove code that causes loops in registrations. 

We have agreed that the patch that this code was part of was bad. I am ripping out the code that causes 
the issue. putnopvut needs to check the rest of the patch, if it needs to be changed as well.

This solves the issue reported in #15540, but needs more work before we close it (as described above).


........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@217669 f38db490-d61c-443f-a65b-d21fe96a405b