fixes segfault in iLBC
For reasons not yet known, it appears possible for an ast_frame
to have a datalen greater than zero while the actual data is NULL
during Packet Loss Concealment.  Most codecs don't support PLC so
this doesn't affect them.  This patch catches the malformed frame
and prevents the crash from occuring.  Additional efforts to determine
why it is possible for a frame to look like this are still being
investigated.

(issue #16979)



git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@228418 f38db490-d61c-443f-a65b-d21fe96a405b