fixes test.c compile issue when TEST_FRAMEWORK is not enabled
The ast_test_status_update() function is defined in test.h.
When TEST_FRAMEWORK is not enabled a macro is defined as a no-op
place holder for this function.  The macro did not contain
the correct number of arguments.  This caused a compile error.

Much thanks to wdoekes for reporting the issue and supplying the
patch!



git-svn-id: http://svn.digium.com/svn/asterisk/trunk@238091 f38db490-d61c-443f-a65b-d21fe96a405b