Blocked revisions 242919 via svnmerge
........
r242919 | oej | 2010-01-25 22:13:20 +0100 (Mån, 25 Jan 2010) | 8 lines

Change api for pbx_builtin_setvar to actually return error code if a function can't be written to.

This patch removes code that was duplicated from pbx.c to manager.c
in order to prevent API change in released versions of Asterisk.

There are propably also other places that would benefit from reading the
return code and react if a function returns error codes on writing a value into it.

........


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.6.2@242920 f38db490-d61c-443f-a65b-d21fe96a405b