Cleanup display_*message functions.
This patch splits transmit_displaymessage into transmit_clear_display_message and transmit_display_message which better aligns with the skinny protocol. The new transmit_display_message is not used in the current code, but will be and so it is commented.

Moved handle_datetime from this function to onhook and offhook functions (display now properly cleared at the end of a call on 30VIP).

Removed skinny debug messages from inline code as there's an ast_verb in transmit_clear_display_message. Also, removed commentary that it was a clear display as it is now apparent from the function name.

Split transmit_displaypromptmessage into display and clear.

(closes issue #16878)
Reported by: wedhorn
Patches: 
	skinny-clean02.diff uploaded by wedhorn (license 30)
	skinny-clean03.diff uploaded by wedhorn (license 30)



git-svn-id: http://svn.digium.com/svn/asterisk/trunk@249670 f38db490-d61c-443f-a65b-d21fe96a405b