Update backtrace.txt documentation.
Update the backtrace.txt documentation so it conforms to the same layout as
other documents we've been working on recently. Additionally, add a bunch of
new information about gathering backtraces for crashes and deadlocks, along
with ways of verifying your file before uploading it. Create a couple of one
line commands for people to generate the files we need.

(closes issue #17190)
Reported by: lmadsen
Patches: 
      backtrace.txt.patch-2 uploaded by lmadsen (license 10)
Tested by: lmadsen, pabelanger

git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@257426 f38db490-d61c-443f-a65b-d21fe96a405b