Commit compromise I suggested on review 608.
This allows for multiple SRV queries to be done
from the dialplan for the same service on a single call while
still allowing one to bypass the call to SRVQUERY if they so
please.

Taking action since no comments had been left for a while.
This can easily be reverted if needed. External tests
still pass.



git-svn-id: http://svn.digium.com/svn/asterisk/trunk@257851 f38db490-d61c-443f-a65b-d21fe96a405b