Remove usage of `id` since it isn't useful and was causing breakge.
Solaris `id` doesn't support the -u argument.  Instead of figuring out how to
fix this to work on Solaris, I decided to check why it was necessary and where
else it was used.  It was only used in one place, and it hasn't been needed
for a very long time (I question whether it was ever needed).


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@259748 f38db490-d61c-443f-a65b-d21fe96a405b