Voicemail transfer to operator should occur immediately, not after main menu.
There were two scenarios in the advanced options that while using the
operator=yes and review=yes options, the transfer occurred only after exiting
the main menu (after sending a reply or leaving a message for an extension).
Now after the audio is processed for the reply or message the transfer occurs
immediately as expected.

ABE-2107
ABE-2108



git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@260923 f38db490-d61c-443f-a65b-d21fe96a405b