VoicemailMain and VMauthenticate, allow escape to the 'a' extension when a single '*' is entered
Where a site uses VoicemailMain(mailbox) the users have to be at their own extension to clear
their voicemail, they have no way of escaping VoicemailMain to allow entry of new boxnumber.

This patch, allows a site to include to 'a' priority in the VoicemailMain context, to allow an escape.

If the 'a' priority doesn't exist in the context that VoicemailMain was called from then it acts as the old behaviour.

  Reported by: alecdavis
  Tested by: alecdavis
  Patch
	 vm_a_extension.diff2.txt uploaded by alecdavis (license 585)

Review: https://reviewboard.asterisk.org/r/489/


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@262005 f38db490-d61c-443f-a65b-d21fe96a405b