Fix infinite loop when loading codec speex
This changes the sample slinear frame data to contain non-zero data so that
translation calculations for speex works when preprocessing and VAD is turned
on. The encoder expects samples to be returned, but when attempted with the
mentioned two options and silent sample frames everything was discarded. 

(closes issue #17240)
Reported by: seandarcy

Review: https://reviewboard.asterisk.org/r/682/


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@267065 f38db490-d61c-443f-a65b-d21fe96a405b